adding additional log analysis capabilities

+++ scripts/compare_logs.py
@@ -1,14 +1,17 @@
 import argparse
 import json
 import os
+import random
 import shlex
 import sqlite3
 import subprocess
+import sys
 import time
 from typing import Any, Callable, Collection, Dict, List, Optional, Sequence, Tuple
 
 from fastdtw import fastdtw as fdtw
 import numpy as np
+#np.set_printoptions(threshold=sys.maxsize)
 from scipy.spatial.distance import sqeuclidean
 
 
@@ -27,12 +30,13 @@ def get_json(log_fn: str) -> List[Dict]:
     if log_fn.endswith(".json"):
         json_fn = log_fn
     if log_fn.endswith(".tlog"):
-        cmd = f"mavlogdump.py --format json {log_fn}"
         json_fn = f"{log_fn}.json"
-        with open(json_fn, 'w') as json_file:
-            print(cmd)
-            subprocess.Popen(shlex.split(cmd), stdout=json_file)
-    time.sleep(1)
+        if not os.path.isfile(json_fn):
+            cmd = f"mavlogdump.py --format json {log_fn}"
+            with open(json_fn, 'w') as json_file:
+                print(cmd)
+                subprocess.Popen(shlex.split(cmd), stdout=json_file)
+            time.sleep(1)
     with open(json_fn, 'r') as json_read:
         print(json_fn)
         for line in json_read:
@@ -44,12 +48,15 @@ def get_json(log_fn: str) -> List[Dict]:
 def pairwise(ts_a: np.array, ts_b: np.array, weights: np.array,
              alg='eros') -> float:
     if alg == 'eros':
-        return eros_similarity(ts_a, ts_b, weights)
+        comp = eros_similarity(ts_a, ts_b, weights)
     elif alg == 'fdtw':
-        return fastdtw(ts_a, ts_b)
+        comp = fastdtw(ts_a, ts_b)
     else:
         raise NotImplementedError
-
+    print(f"ts_a: {ts_a}")
+    print(f"ts_b: {ts_b}")
+    print(f"alg: {alg} comp: {comp}")
+    return comp
 
 def fastdtw(x: np.array, y: np.array) -> float:
     return np.sqrt(fdtw(x, y, dist=sqeuclidean)[0])
@@ -163,21 +170,36 @@ def eros_weights(dataset: Collection[np.array],
 def access_bag_db(db_fn: str) -> Tuple[sqlite3.Cursor, sqlite3.Connection]:
     conn = sqlite3.connect(db_fn)
     if conn is not None:
-        c = conn.cursor()
+        cursor = conn.cursor()
     else:
         raise("Error! Cannot create database connection!")
-    return c, conn
+    return cursor, conn
+
 
 def get_fns_from_rows(cursor: sqlite3.Cursor, log_dir: str) -> List[str]:
     rows = cursor.fetchall()
     log_fns = []
     for row in rows:
-        print(row)
+        # print(row)
         log_fn = os.path.join(log_dir, f"{row[0]}.tlog")
-        print(log_fn)
+        # print(log_fn)
         log_fns.append(log_fn)
     return log_fns
 
+
+def memoize_dst(log_fn_1, log_fn_2, db_fn):
+    # Access the database
+    cursor, conn = access_bag_db(db_fn)
+
+    # Check if the database has the correct table.
+    # TODO
+    # If not, create the correct table.
+    # TODO
+    # Check if the distance is in the table
+    # TODO
+    # If so return it
+    # TODO
+    # If not, calculate it
 
 def get_from_db(log_db: str, log_dir: str = "../bags") -> Dict[str, List[str]]:
     log_fns: Dict[str, List[np.array]] = dict()
@@ -187,10 +209,12 @@ def get_from_db(log_db: str, log_dir: str = "../bags") -> Dict[str, List[str]]:
     cursor.execute("select * from bagfns where mutation_fn=?", (filename,))
     nominal_fns = get_fns_from_rows(cursor, log_dir)
     log_fns['nominal'] = nominal_fns
+    print(f"len(nominal_fns): {len(nominal_fns)}")
+    print(f"len(log_fns['nominal']): {len(log_fns['nominal'])}")
+
     cursor.execute("select * from bagfns where mutation_fn!=?", (filename,))
     experimental_fns = get_fns_from_rows(cursor, log_dir)
     log_fns['experimental'] = experimental_fns
-
 
     conn.close()
     return log_fns
@@ -198,126 +222,130 @@ def get_from_db(log_db: str, log_dir: str = "../bags") -> Dict[str, List[str]]:
 
 def convert_logs(log_fns: List[str]) -> List[np.array]:
     global_pos_lists = []
+    fn_count = 0
+    total_fns = len(log_fns)
     for log_fn in log_fns:
         # convert the tlog to json, if it's not already json
         log_json = get_json(log_fn)
-        #print(log_json)
         global_pos = [x["data"] for x in log_json if x["meta"]["type"] == "GLOBAL_POSITION_INT"]
         global_pos_list = ([(x['time_boot_ms'], x['lat'], x['lon'],
                              x['alt'], x['relative_alt'], x['vx'],
                              x['vy'], x['vz'], x['hdg']) for x in
                             global_pos])
-        #print(global_pos_list)
         global_np = np.array(global_pos_list)
         global_pos_lists.append(global_np)
-
+        fn_count += 1
+        print(f"File {fn_count} of {total_fns}")
     return global_pos_lists
 
 
-def get_logs(args: argparse.Namespace)-> Tuple[List[np.array], List[str]]:
-    log_fns: List[np.array] = list()
-    labels = list()
+def insert(db: Dict[str, List[Any]], label: str, data: List[Any]) -> Dict[str, List[Any]]:
+    print(db)
+    if label in db:
+        db[label] = db[label] + data
+    else:
+        db[label] = data
+    return db
+
+
+def get_logs(args: argparse.Namespace)-> Dict[str, List[np.array]]:
+    log_fns: Dict[str, List[np.array]] = dict()
 
     if args.log_fn and len(args.log_fn) > 1:
-        log_fns.extend(args.log_fn)
-        labels.extend(['manual'] * len(args.log_fn))
+        log_fns = insert(log_fns, "manual", args.log_fn)
     elif args.log_db:
-        logs = get_from_db(args.log_db, args.log_dir)
-        for label in logs:
-            log_fns.extend(logs[label])
-            labels.extend([label] * len(logs[label]))
+        log_fns = get_from_db(args.log_db)
 
+    all_logs = dict()
     # Turn the log files into np.arrays of the data
-    # TODO
-    logs_data = convert_logs(log_fns)
+    for label, log_fns_subset in log_fns.items():
+        logs_data = convert_logs(log_fns_subset)
+        all_logs[label] = logs_data
 
-    return logs_data, labels
+    print(f"all_logs: {all_logs}")
+    return all_logs
 
 
 def get_comparisons(logs: List[np.array],
                     weights: np.array = None, alg: str = 'eros') -> np.array:
-    comparisons = np.zeros(len(logs), len(logs))
+    comparisons = np.zeros((len(logs), len(logs)))
     for i in range(len(logs)):
         for j in range(len(logs)):
-            comparison = pairwise(logs[i], logs[j], weights=weights,
-                                  alg=alg)
-        comparisons[i][j] = comparison
+            if comparisons[i][j] == 0:
+                print(f"comparing log[{i}] and log[{j}]")
+                comparison = pairwise(logs[i], logs[j], weights=weights,
+                                      alg=alg)
+                comparisons[i][j] = comparison
+                comparisons[j][i] = comparison
     return comparisons
 
 
-def compare_logs(logs: List[np.array], labels: List[str],
+def equalize_lengths(logs: List[np.array], length: int=0) -> List[np.array]:
+    if length == 0:
+        length = min([len(x) for x in logs])
+
+    new_logs = []
+    for log in logs:
+        while len(log) > length:
+            to_delete = random.randrange(len(log))
+            log = np.delete(log, to_delete, axis=0)
+        new_logs.append(log)
+
+    #print(str([x for x in new_logs]))
+    return new_logs
+
+def compare_logs(logs_dict: Dict[str, List[np.array]],
                  alg: str = 'eros') -> Dict[str, np.array]:
-    assert(np.size(0,) == len(labels))
-    label_logs = zip(labels, logs)
+    #assert(len(logs) == len(labels)), f"{len(logs)} {len(labels)}"
+    #label_logs = zip(labels, logs)
 
     # Get all the groups for individual comparisons, by labels
-    label_set = set(labels)
-    logs_dict = dict()
-    for label in labels:
-        logs_subset = [x[1] for x in label_logs if x[0] == label]
-        logs_dict[label] = logs_subset
+    #label_set = set(labels)
+    #logs_dict = dict()
+    #for label in label_set:
+    #    logs_subset = [x[1] for x in label_logs if x[0] == label]
+        # print(f"label: {label} logs_subset: {logs_subset}")
+    #    logs_dict[label] = logs_subset
+    #    print(f"len(logs_dict[{label}]): {len(logs_dict[label])}")
+
+    print(f"logs_dict.keys(): {logs_dict.keys()}")
+
+    all_logs = []
+    for label, logs in logs_dict.items():
+        all_logs.extend(logs)
 
     if alg == 'eros':
-        n = len(logs[0])
-        assert(all([len(x) == n for x in logs]))
-        weights = eros_weights(logs)
+        n = len(all_logs[0])
+        # print(logs[0])
+        if not all([len(x) == n for x in all_logs]):
+            logs_eq = equalize_lengths(all_logs)
+        else:
+            logs_eq = all_logs
+        weights = eros_weights(logs_eq)
     else:
         weights = None
 
+    print(f"logs_dict.keys(): {logs_dict.keys()}")
     all_comparisons = dict()
-    for label, logs_subset in (list(logs_dict.items()) +
-                               list([('total', logs)])):
+    for label, logs_subset in (list(logs_dict.items())): # +
+                               #list([('total', logs)])):
+        print(f"label: {label} len(logs_subset): {len(logs_subset)}")
         if len(logs_subset) > 1:
             comparisons = get_comparisons(logs, weights, alg)
             all_comparisons[label] = comparisons
 
     return all_comparisons
 
+
 def main():
     args = parse_args()
 
     # get the logs to compare
-    logs, labels = get_logs(args)
+    logs_dict = get_logs(args)
 
     # compare the logs
-    comparisons = compare_logs(logs, labels)
-
-    """
-    global_pos_lists = []
-
-    if args.log_fn and len(args.log_fn) > 0:
-        log_fns = args.log_fn
-    elif args.use_nominal:
-        nominal_log_fns = get_nominal(args.log_db)
-        log_fns = nominal_log_fns
-
-    for log_fn in log_fns:
-        # convert the tlog to json, if it's not already json
-        log_json = get_json(log_fn)
-        #print(log_json)
-        global_pos = [x["data"] for x in log_json if x["meta"]["type"] == "GLOBAL_POSITION_INT"]
-        global_pos_list = ([(x['time_boot_ms'], x['lat'], x['lon'],
-                             x['alt'], x['relative_alt'], x['vx'],
-                             x['vy'], x['vz'], x['hdg']) for x in
-                            global_pos])
-        #print(global_pos_list)
-        global_np = np.array(global_pos_list)
-        global_pos_lists.append(global_np)
-
-    comparisons = np.zeros((len(global_pos_lists), len(global_pos_lists)))
-    print(comparisons.shape)
-    weights = eros_weights(global_pos_lists)
-    if len(comparisons) > 1:
-        for i in range(len(global_pos_lists)):
-            for j in range(len(global_pos_lists)):
-                comparison = pairwise(global_pos_lists[i],
-                                      global_pos_lists[j], weights,
-                                      alg=args.alg)
-                #print(comparison)
-                comparisons[i][j] = comparison
-
-    print(comparisons)
-"""
+    comparisons = compare_logs(logs_dict, args.alg)
+    print(f"comparisons: {comparisons}")
 
 
 if __name__ == '__main__':
